Adding Tableau calculations
As a stop gap for #5 to make some aggregate data publicly available calculate statistics in a Tableau aggregate extract. Eventually this will have to be pushed down to the underlying database, for performance and for better accuracy in the calculation of standard errors; which currently does not account for within person correlations between the shortest and longest lifespan rates.